requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.3
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.0.0
pytesseract==0.3.10
openpyxl==3.1.2